Cast guest_ram_mb to prevent signed overflow
guest_ram_mb is declared by camkes as a signed `int`, and if too large will overflow
into a negative number, prior to conversion to being a `size_t`, which is the final
type of the argument

+++ components/Init/src/main.c
@@ -637,7 +637,7 @@ void *main_continued(void *arg) {
 #else
     paddr_is_vaddr = 0;
 #endif
-    error = vmm_alloc_guest_ram(&vmm, guest_ram_mb * 1024 * 1024, paddr_is_vaddr);
+    error = vmm_alloc_guest_ram(&vmm, (size_t)guest_ram_mb * 1024 * 1024, paddr_is_vaddr);
     ZF_LOGF_IF(error, "Failed to allocate guest ram");
 
     /* Perform device discovery and give passthrough device information */
